+++ demos/59-circle-and-ellipse-tangents/index.js
@@ -26,41 +26,35 @@
         GUP
       )
     );
-    // Let's set up some colors.
-    pb.drawConfig.polygon.color = "rgba(255,192,0,0.75)";
-    pb.drawConfig.polygon.lineWidth = 2;
 
     // Circle
-    var circle = null;
-    var ray = new Vector(new Vertex(), new Vertex(100, 100));
+    var circle = new Circle(new Vertex(), 90.0);
+    var ellipse = new VEllipse(new Vertex(), new Vertex(200, 150), Math.PI / 3.0);
+    // ellipse.axis.attr.visible = false;
+    // ellipse.axis.attr.draggable = false;
+
+    // We want to change the ellipse's radii and rotation by dragging points around
+    var ellipseRotationControlPoint = ellipse.vertAt(0.0).scale(1.2, ellipse.center);
+
+    var ellipseHelper = new VEllipseHelper(ellipse, ellipseRotationControlPoint);
+    pb.add([circle, ellipse, ellipseRotationControlPoint], false);
 
     // Create a config: we want to have control about the arrow head size in this demo
     var config = {
-      angle: params.getNumber("angle", 0.0)
+      angle: params.getNumber("angle", Math.round(ellipse.rotation * RAD_TO_DEG))
     };
 
     // +---------------------------------------------------------------------------------
     // | Global vars
     // +-------------------------------
-    var viewport = pb.viewport();
-    // Must be in clockwise order!
-    var polygon = null;
+    // var viewport = pb.viewport();
 
     var postDraw = function (draw, fill) {
-      var tangent = circle.tangentAt(config.angle * DEG_TO_RAD);
-      draw.arrow(tangent.a, tangent.b, "orange");
-    };
-
-    // +---------------------------------------------------------------------------------
-    // | Just rebuilds the pattern on changes.
-    // +-------------------------------
-    var rebuildShape = function () {
-      // Re-add the new polygon to trigger redraw.
-      pb.removeAll(false, false); // Don't trigger redraw
-
-      circle = new Circle(new Vertex(), 90.0);
-      pb.add(circle, true);
-      // pb.add(ray, true);
+      var circleTangent = circle.tangentAt(config.angle * DEG_TO_RAD);
+      draw.arrow(circleTangent.a, circleTangent.b, "orange", 2.0);
+      var ellipseTangent = ellipse.tangentAt(config.angle * DEG_TO_RAD);
+      draw.arrow(ellipseTangent.a, ellipseTangent.b, "green", 2.0);
+      ellipseHelper.drawHandleLines(draw, fill);
     };
 
     // +---------------------------------------------------------------------------------
@@ -74,6 +68,6 @@
     }
 
     pb.config.postDraw = postDraw;
-    rebuildShape();
+    pb.redraw();
   });
 })(globalThis);